+++ api/stdcopy/stdcopy_test.go
