+++ integration/capabilities/capabilities_linux_test.go
@@ -7,12 +7,11 @@ import (
 	"testing"
 
 	"github.com/docker/docker/integration/internal/container"
-	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/docker/docker/testutil"
 	"github.com/docker/docker/testutil/fakecontext"
+	"github.com/moby/moby/api/stdcopy"
 	"github.com/moby/moby/api/types/build"
 	containertypes "github.com/moby/moby/api/types/container"
-
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/poll"
 )
