+++ integration/plugin/logging/read_test.go
@@ -8,9 +8,9 @@ import (
 	"time"
 
 	testContainer "github.com/docker/docker/integration/internal/container"
-	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/docker/docker/testutil"
 	"github.com/docker/docker/testutil/daemon"
+	"github.com/moby/moby/api/stdcopy"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/container"
 	"gotest.tools/v3/assert"
